+++ CHANGELOG.md
@@ -4,6 +4,7 @@ GAMSPy CHANGELOG
 GAMSPy 1.7.0
 ------------
 - General
+  - Allow container serialization/deserialization.
   - Support an alternative syntax for operations. For example, x.sum() is equivalent to Sum(x.domain, x[x.domain]).
   - Fix a bug when starting from a GAMS restart file.
   - Allow propagating bounds of the output in `Conv2D` class.